Show how to leverage React Server Components more
