+++ src/app/products/[productId]/loading.tsx
@@ -1,0 +1,3 @@
+export default function ProductLoading() {
+    return <div>loading...</div>
+}
