--- src/app/products/services/getAllProduct.ts
@@ -1,10 +1,0 @@
-import { getBaseUrl } from "@/lib/getBaseURL";
-
-export async function getAllProduct() {
-  const response = await fetch(`${getBaseUrl()}/api/products/get-all`, {
-    next: { revalidate: 0 },
-  });
-  if (!response.ok) throw new Error("Failed to fetch all products");
-
-  return response.json();
-}
